Move processing app list to module
