+++ lib/search.js
@@ -1,11 +1,9 @@
 'use strict';
 
 const request = require('./utils/request');
-// const cheerio = require('cheerio');
-const matchScriptData = require('./utils/matchScriptData');
-const generateExtractFieldsFn = require('./utils/generateExtractFieldsFn');
-const R = require('ramda');
-const url = require('url');
+const scriptData = require('./utils/scriptData');
+const mapping = require('./utils/appListMapping');
+
 // const queryString = require('querystring');
 
 // function getClp(html) {
@@ -34,42 +32,10 @@ const url = require('url');
  * `checkFinished` to repeat the process with next page if necessary.
  */
 function processAndRecur (html, opts, savedApps, clp) {
-  const data = matchScriptData(html);
-  const input = R.path(['ds:3', 0, 1, 0, 0, 0], data);
-  const apps = R.map(generateExtractFieldsFn(MAPPINGS), input);
+  const data = scriptData.parse(html);
+  const apps = mapping.processList(data);
 
   return checkFinished(opts, [...savedApps, ...apps], null, clp);
-}
-
-const MAPPINGS = {
-  title: [2],
-  appId: [12, 0],
-  url: {
-    path: [9, 4, 2],
-    fun: (path) => new url.URL(path, 'https://play.google.com').toString()
-  },
-  icon: [1, 1, 0, 3, 2],
-  developer: [4, 0, 0, 0],
-  developerId: {
-    path: [4, 0, 0, 1, 4, 2],
-    fun: extaractDeveloperId
-  },
-  priceText: {
-    path: [7, 0, 3, 2, 1, 0, 2],
-    fun: (price) => price === undefined ? 'FREE' : price
-  },
-  free: {
-    path: [7, 0, 3, 2, 1, 0, 2],
-    fun: (price) => price === undefined
-  },
-  summary: [4, 1, 1, 1, 1],
-  scoreText: [6, 0, 2, 1, 0],
-  score: [6, 0, 2, 1, 1]
-};
-
-function extaractDeveloperId (link) {
-  const q = url.parse(link, {parseQueryString: true});
-  return q.query['id'];
 }
 
 /*
