+++ lib/similar.js
@@ -1,10 +1,9 @@
 'use strict';
 
 const request = require('./utils/request');
-const R = require('ramda');
-const url = require('url');
 const queryString = require('querystring');
 const scriptData = require('./utils/scriptData');
+const mapping = require('./utils/appListMapping');
 
 const PLAYSTORE_URL = 'https://play.google.com/store/apps/details';
 
@@ -32,7 +31,7 @@ function similar (opts) {
 
     request(options, opts.throttle)
       .then(scriptData.parse)
-      .then(scriptData.extractor(PATH_MAPPING))
+      .then(scriptData.extractor(PATH))
       .then(data => {
         if (data.path === undefined) {
           throw Error('not found similar link');
@@ -43,55 +42,14 @@ function similar (opts) {
         }, opts.requestOptions));
       })
       .then(scriptData.parse)
-      .then(data => {
-        const input = R.path(['ds:3', 0, 1, 0, 0, 0], data);
-        return R.map(scriptData.extractor(MAPPINGS), input);
-      })
+      .then(mapping.processList)
       .then(resolve)
       .catch(reject);
   });
 }
 
-const PATH_MAPPING = {
+const PATH = {
   path: ['ds:6', 1, 1, 0, 0, 3, 4, 2]
 };
-
-/*
- * Map the MAPPINGS object, applying each field spec to the parsed data.
- * If the mapping value is an array, use it as the path to the extract the
- * field's value. If it's an object, extract the value in object.path and pass
- * it to the function in object.fun
-*/
-
-const MAPPINGS = {
-  title: [2],
-  appId: [12, 0],
-  url: {
-    path: [9, 4, 2],
-    fun: (path) => new url.URL(path, 'https://play.google.com').toString()
-  },
-  icon: [1, 1, 0, 3, 2],
-  developer: [4, 0, 0, 0],
-  developerId: {
-    path: [4, 0, 0, 1, 4, 2],
-    fun: extractDeveloperId
-  },
-  priceText: {
-    path: [7, 0, 3, 2, 1, 0, 2],
-    fun: (price) => price === undefined ? 'FREE' : price
-  },
-  free: {
-    path: [7, 0, 3, 2, 1, 0, 2],
-    fun: (price) => price === undefined
-  },
-  summary: [4, 1, 1, 1, 1],
-  scoreText: [6, 0, 2, 1, 0],
-  score: [6, 0, 2, 1, 1]
-};
-
-function extractDeveloperId (link) {
-  const q = url.parse(link, { parseQueryString: true });
-  return q.query['id'];
-}
 
 module.exports = similar;
