+++ lib/utils/appListMapping.js
@@ -1,0 +1,50 @@
+'use strict';
+
+const url = require('url');
+const R = require('ramda');
+const scriptData = require('./scriptData');
+
+/*
+ * Map the MAPPINGS object, applying each field spec to the parsed data.
+ * If the mapping value is an array, use it as the path to the extract the
+ * field's value. If it's an object, extract the value in object.path and pass
+ * it to the function in object.fun
+*/
+
+const MAPPINGS = {
+  title: [2],
+  appId: [12, 0],
+  url: {
+    path: [9, 4, 2],
+    fun: (path) => new url.URL(path, 'https://play.google.com').toString()
+  },
+  icon: [1, 1, 0, 3, 2],
+  developer: [4, 0, 0, 0],
+  developerId: {
+    path: [4, 0, 0, 1, 4, 2],
+    fun: extaractDeveloperId
+  },
+  priceText: {
+    path: [7, 0, 3, 2, 1, 0, 2],
+    fun: (price) => price === undefined ? 'FREE' : price
+  },
+  free: {
+    path: [7, 0, 3, 2, 1, 0, 2],
+    fun: (price) => price === undefined
+  },
+  summary: [4, 1, 1, 1, 1],
+  scoreText: [6, 0, 2, 1, 0],
+  score: [6, 0, 2, 1, 1]
+};
+
+function extaractDeveloperId (link) {
+  const q = url.parse(link, {parseQueryString: true});
+  return q.query['id'];
+}
+
+function processList (data) {
+  const input = R.path(['ds:3', 0, 1, 0, 0, 0], data);
+  return R.map(scriptData.extractor(MAPPINGS), input);
+}
+
+module.exports = Object.assign({MAPPINGS, processList});
